fms/requirements.txt
Authlib==1.6.3
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
Flask==3.1.0
Flask-Session==0.8.0
fonttools==4.58.5
future==1.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
msgspec==0.19.0
netmiko==4.5.0
ntc_templates==7.8.0
numpy==2.3.1
packaging==25.0
paramiko==3.5.1
pillow==11.3.0
pycparser==2.22
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
schedule==1.2.2
scp==0.15.0
six==1.17.0
textfsm==1.1.3
urllib3==2.4.0
Werkzeug==3.1.3
